LPS-116453 Splitting into 2 separate interfaces
