+++ modules/apps/translation/translation-api/src/main/resources/com/liferay/translation/importer/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
